refactor: Split ModelsExtensions.kt into per-type dedicated files
